+++ CHANGELOG.md
@@ -27,6 +27,7 @@ omnibus-gitlab repository.
 - Update postgresql to 9.2.18 (Takuya Noguchi)
 - Update expat to 2.2.0 (Takuya Noguchi)
 - Ignore and don't write `gitlab_ci:gitlab_server` key in gitlab-secrets file 10bcb
+- Configure Rake attack (Dmitry Ivanov)
 
 8.10.7